# perf: a faster implementation of Hex()