While I was benchmarking my bubble tea application, I noticed that it was spending almost 40% of its time in go-colorful's Hex function, and specifically in the fmt.Sscanf function.

This PR implements Hex() without relying on fmt.Sscanf, but instead performs its own parsing. All tests pass.

The performance improvement is substantial:

```
before: BenchmarkHex-10    	 3761415	       319.4 ns/op	      74 B/op	       6 allocs/op
after:  BenchmarkHex-10    	90607219	       12.50 ns/op	       0 B/op	       0 allocs/op
```

Not only is it significantly faster. It also does not perform any allocations. 

The impact on my own application is also significant: in the current implementation, it spent about 40% of the time in Hex().  Now this is down to less than 5%:

```
before: BenchmarkTable_View/styled-10         	     921	   1433068 ns/op	  272142 B/op	   11766 allocs/op
after:  BenchmarkTable_View/styled-10         	    1525	    782587 ns/op	   75496 B/op	    1566 allocs/op
```

Comments are welcome! :)