@Encephala 2025-12-14
Sorry to reply after the PR was merged, I couldn't find a more logical place to put this.

I just found myself quite confused why updating the dependencies in my Go TUI project made colours stop working. Weeding through the dependencies, I saw colorful updated, and then I stumbled on this PR.

Turns out I was using RGBA hex codes (f.i. #F6D6D680 for semi-transparent salmon pink), and the old parser was parsing the RGB out of those just fine, ignoring the A part. The new parser does not parse these colours succesfully, I suppose because of [this switch statement](https://github.com/clambin/go-colorful/blob/7762334a86980f349239f4a2d1a976983d93d1b4/colors.go#L373).

I'm not sure whether you would want to consider this a breaking change. I'm not even sure whether any terminals support transparent colours, nor am I sure why I put alpha transparency into my hex codes.
But perhaps it is worthwhile to put a note on this in the changelog/the github release?